antialiasing

+++ shape/Circle.h
@@ -6,13 +6,13 @@
 class CCircle : public ISolidShape
 {
 public:
-    CCircle(Point const& center, GLfloat radius, Color const& outlineColor, Color const& fillColor);
+    CCircle(Point const& center, float radius, Color const& outlineColor, Color const& fillColor);
     ~CCircle() = default;
-    double GetArea() const override;
-    double GetPerimeter() const override;
+    float GetArea() const override;
+    float GetPerimeter() const override;
     Color GetOutlineColor() const override;
     Color GetFillColor() const override;
-    double GetRadius() const;
+    float GetRadius() const;
     Point const& GetCenter() const;
 
     void Draw(ICanvas & canvas) const override;
@@ -20,7 +20,7 @@ protected:
     void AppendProperties(std::ostream & strm) const override;
 private:
     Point m_center;
-    GLfloat m_radius;
+    float m_radius;
     Color m_outlineColor;
     Color m_fillColor;
 };
